* gc.c (STACK_LEVEL_MAX, ruby_stack_length): returns size_t.   [ruby-core:18207]
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@18463 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/intern.h
@@ -330,7 +330,7 @@ VALUE rb_file_directory_p(VALUE,VALUE);
 void ruby_set_stack_size(size_t);
 NORETURN(void rb_memerror(void));
 int ruby_stack_check(void);
-int ruby_stack_length(VALUE**);
+size_t ruby_stack_length(VALUE**);
 int rb_during_gc(void);
 void rb_gc_mark_locations(VALUE*, VALUE*);
 void rb_mark_tbl(struct st_table*);
